create asteroid box

+++ Models/Asteroid.cs
@@ -9,7 +9,7 @@
         public string Hazardous { get; set; }
         public string Time { get; set; }
         public decimal RelativeVelocity { get; set; }
-        public decimal MissDistance { get; set; }
+        public int MissDistance { get; set; }
         public string Date { get; set; }
     }
 }
